Overføring av medlemmer fra skrift.no ferdig
